DB-11755 Use int for sketchsize

+++ db-engine/src/main/java/com/splicemachine/db/iapi/sql/dictionary/ColumnDescriptor.java
@@ -81,7 +81,7 @@ public final class ColumnDescriptor extends TupleDescriptor
     private boolean collectStatistics;
     private int partitionPosition = -1;
     private byte  useExtrapolation = 0;
-    private long sketchSize = 0;
+    private int sketchSize = 0;
     /* Used for Serde */
     //Following variable is used to see if the user is adding an autoincrement
     //column, or if user is altering the existing autoincrement column to change
@@ -289,7 +289,7 @@ public final class ColumnDescriptor extends TupleDescriptor
         return useExtrapolation;
     }
 
-    public long getSketchSize() { return sketchSize; }
+    public int getSketchSize() { return sketchSize; }
 
     public void setUseExtrapolation(byte useExtrapolation) {
         this.useExtrapolation = useExtrapolation;
